requirements.txt
opencv-python 
ultralytics 
pillow 
torch 
numpy
